backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==44.0.3
Django==5.2.1
django-extensions==4.1
djangorestframework==3.16.0
idna==3.10
MarkupSafe==3.0.2
pycparser==2.22
pyOpenSSL==25.0.0
python-dotenv==1.1.0
requests==2.32.3
sqlparse==0.5.3
urllib3==2.4.0
Werkzeug==3.1.3
